.vimrc: quickfix colorscheme
after the latest update, my color scheme
in vim was off. i found the 'correct' colors
were used when i set the background to light.

This indicates issues with my color scheme setup
in vim, but i cannot deal with that right now.
instead, set the default background color in vim
to 'light'.

Signed-off-by: einKnie <einKnie@gmx.at>