i3/config_legion: add screenshot mode
add mode for screenshots, can either take all displays
or an interactive selection

This uses maim instead of scrot. not relevant for this, but i've run
into issues running a bindsym with scrot and autoselection (for
multiscreen setup) and maim works a lot better in this regard.
therefore -> switching to maim generally

Signed-off-by: einKnie <einKnie@gmx.at>